requirements.txt
# LiveKit Agent Dependencies with MCP support
livekit-agents[codecs,images,openai,silero,mcp]
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-noise-cancellation

# Environment and Configuration
python-dotenv

# Additional dependencies that might be needed
aiohttp
asyncio

# Google Calendar API dependencies
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# DateTime tools dependencies
pytz